+++ src/extension.ts
@@ -4,12 +4,16 @@ import { activate as activateAPI } from './api';
 import { activate as activateAutocomplete } from './autocomplete';
 import { activate as activateBuild } from './build';
 import { activate as activateDebug } from './debug';
+import { activate as activateDefinition } from './definition';
 import { activate as activateDiagnostic } from './diagnostic';
 import { activate as activateHover } from './hover';
 import { activate as activateNextError } from './next-error';
 import { activate as activateRefresh } from './refresh';
 import { activate as activateSubscriptions } from './subscriptions';
+import { activate as activateSymbol } from './symbol';
 import { activate as activateTransform } from './transform';
+import { activate as activateSnippets } from './snippet';
+import { activate as activateColor } from './color';
 
 export function activate(context: ExtensionContext) {
   const config = vscode.workspace.getConfiguration('greybel');
@@ -29,12 +33,16 @@ export function activate(context: ExtensionContext) {
   activateBuild(context);
   activateTransform(context);
   activateNextError(context);
+  activateDefinition(context);
+  activateSymbol(context);
 
   if (config.get<boolean>('diagnostic')) {
     activateDiagnostic(context);
   }
 
   activateAPI(context);
+  activateSnippets(context);
+  activateColor(context);
 }
 
 export function deactivate() {}