+++ src/helper/document-manager.ts
@@ -1,9 +1,12 @@
 import EventEmitter from 'events';
 import { ASTChunkAdvanced, Parser } from 'greybel-core';
 import { ASTBase } from 'greyscript-core';
-import { TextDocument } from 'vscode';
+import LRU from 'lru-cache';
+import vscode, { TextDocument } from 'vscode';
 
 export interface ParseResult {
+  content: string;
+  textDocument: TextDocument;
   document: ASTBase | null;
   errors: Error[];
 }
@@ -14,10 +17,10 @@ export interface QueueItem {
 }
 
 export const DOCUMENT_PARSE_QUEUE_INTERVAL = 1000;
-export const DOCUMENT_PARSE_QUEUE_PARSE_TIMEOUT = 5000;
+export const DOCUMENT_PARSE_QUEUE_PARSE_TIMEOUT = 2500;
 
 export class DocumentParseQueue extends EventEmitter {
-  results: Map<string, ParseResult>;
+  results: LRU<string, ParseResult>;
 
   private queue: Map<string, QueueItem>;
   private interval: NodeJS.Timeout | null;
@@ -25,7 +28,10 @@ export class DocumentParseQueue extends EventEmitter {
 
   constructor(parseTimeout: number = DOCUMENT_PARSE_QUEUE_PARSE_TIMEOUT) {
     super();
-    this.results = new Map();
+    this.results = new LRU({
+      ttl: 1000 * 60 * 20,
+      ttlAutopurge: true
+    });
     this.queue = new Map();
     this.interval = null;
     this.parseTimeout = parseTimeout;
@@ -67,6 +73,7 @@ export class DocumentParseQueue extends EventEmitter {
     this.results.set(key, result);
     this.emit('parsed', document, result);
     this.queue.delete(key);
+
     return result;
   }
 
@@ -79,6 +86,8 @@ export class DocumentParseQueue extends EventEmitter {
 
     if ((chunk as ASTChunkAdvanced).body?.length > 0) {
       return {
+        content,
+        textDocument: document,
         document: chunk,
         errors: parser.errors
       };
@@ -89,11 +98,15 @@ export class DocumentParseQueue extends EventEmitter {
       const strictChunk = strictParser.parseChunk();
 
       return {
+        content,
+        textDocument: document,
         document: strictChunk,
         errors: []
       };
     } catch (err: any) {
       return {
+        content,
+        textDocument: document,
         document: null,
         errors: [err]
       };
@@ -102,8 +115,13 @@ export class DocumentParseQueue extends EventEmitter {
 
   update(document: TextDocument): boolean {
     const fileName = document.fileName;
+    const content = document.getText();
 
     if (this.queue.has(fileName)) {
+      return false;
+    }
+
+    if (this.results.get(fileName)?.content === content) {
       return false;
     }
 
@@ -117,8 +135,38 @@ export class DocumentParseQueue extends EventEmitter {
     return true;
   }
 
+  async open(target: string): Promise<ParseResult> {
+    const textDocument = await vscode.workspace.openTextDocument(target);
+    return this.get(textDocument);
+  }
+
   get(document: TextDocument): ParseResult {
     return this.results.get(document.fileName) || this.refresh(document);
+  }
+
+  next(document: TextDocument, timeout: number = 5000): Promise<ParseResult> {
+    const me = this;
+
+    if (me.queue.has(document.fileName)) {
+      return new Promise((resolve) => {
+        const onTimeout = () => {
+          me.removeListener('parsed', onParse);
+          resolve(me.get(document));
+        };
+        const onParse = (evDocument: TextDocument) => {
+          if (evDocument.fileName === document.fileName) {
+            me.removeListener('parsed', onParse);
+            clearTimeout(timer);
+            resolve(me.get(document));
+          }
+        };
+        const timer = setTimeout(onTimeout, timeout);
+
+        me.addListener('parsed', onParse);
+      });
+    }
+
+    return Promise.resolve(me.get(document));
   }
 
   clear(document: TextDocument): void {